Added some rendered animations for testing Made an Frames class and AnimatedController class for loading and playing frames Added Orange to the global colors list Setup test code for changing animations based on user input in Game FIX: aligned_ptr::Fill was taking int, change to const T & Changed the params for Sprite::Draw to take floats instead of ints.

+++ Engine/Sprite.h
@@ -9,7 +9,7 @@ public:
 	Sprite(const std::string &Filename, const WicInitializer &Wic);
 	virtual~Sprite();
 
-	virtual void Draw( int OffsetX, int OffsetY, Graphics &Gfx )const;
+	virtual void Draw( float OffsetX, float OffsetY, Graphics &Gfx )const;
 
 	int GetWidth()const;
 	int GetHeight()const;
